Move Cotengra impl to numerics for the optimizer_factory
