+++ src/numerics/cotengra/manifest.json
